version 0.3 changes
Lots of little fixes, but mostly the addition of handling Stereo 8 bit and 16 bit samples!  This is a game changer for me, but those are standard XM features that should be supported.  I personally hate macros most of the time.  So I removed the Linear Interpolation and Ramping macros.  Also fixed other nasty audio glitches.

Still I am not fully satisfied with this release as I find there is still too much little clicky moments in some sounds that don't start perfectly for example.  Other thing I don't like is the quality of the Linear Interpolation built in this which really is a low pass filter at roughly 8khz, and it sounds horrible...  So by default is disabled.